split off arima models into new module, so it doesn't need to be loaded by problems module
